src/test/java/com/hrm/qa/testcases/HomePageTest.java
package com.hrm.qa.testcases;

import org.testng.Assert;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

import com.hrm.qa.base.TestBase;
import com.hrm.qa.pages.ContactsPage;
import com.hrm.qa.pages.HomePage;
import com.hrm.qa.pages.LoginPage;

public class HomePageTest extends TestBase {
	
	LoginPage loginPage;
	HomePage homePage;
	ContactsPage contactsPage;
	
	public HomePageTest() {
		super();
	}
	
	@BeforeMethod
	public void setUp() {
		
		initialization();
		loginPage = new LoginPage();
		homePage = loginPage.login(prop.getProperty("username"), prop.getProperty("password"));
		
	}
	
	@Test (priority = 1)
	public void verifyHomePageTitleTest() {
		String homePageTitle = homePage.verifyHomePageTitle();
		Assert.assertEquals(homePageTitle, "Cogmento CRM");
	}
	
	@Test (priority = 2)
	public void verifyCorrectUserNameTest() {
	
		Assert.assertTrue(homePage.correctUserName(), "User Name NOT matched");
	}
	
	@Test (priority = 3)
	public void verifyConatctsLinkTest() {
		
		contactsPage = homePage.clickOnContactsLink();
		
	}
	
		
	@AfterMethod
	public void tearDown() {
		//driver.quit();
		
	}

}
